--- reaction/src/App.js
@@ -1,48 +1,0 @@
-import logo from './logo.svg';
-import './App.css';
-import React from 'react';
-
-class Panel extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = { start_time: 0, ran_once: false, counting: false, true_duration: 0, reaction_time: 0, color: 'green'};
-    this.process_click = this.process_click.bind(this);
-  }
-  handle_color = (c) => {
-    // TODO: Your code here!
-  }
-  start_count() {
-    // TODO: Your code here!
-  }
-  end_count() {
-    // TODO: Your code here!
-  }
-  process_click() {
-    if (this.state.counting) {
-      this.end_count();
-    } else this.start_count();
-  }
-  render() {
-    let msg = "Hello World!";
-    // TODO: Your code here!
-    return (
-      <div className = "PanelContainer" onClick = {this.process_click} style={ { background: this.state.color} }>
-        <div className = "Panel">{msg}</div>
-      </div>
-    );
-  }
-}
-
-function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <h1 className =  "Header">How Fast is your Reaction Time?</h1>
-        <Panel />
-        <p>Click as soon as the red box turns green. Click anywhere in the box to start.</p>
-      </header>
-    </div>
-  );
-}
-
-export default App;
